+++ test/ts-demo/src/car.ts
@@ -1,0 +1,9 @@
+interface Driveable {
+	drive(distance: number): boolean;
+}
+
+export default class Car implements Driveable {
+	public drive(distance: number): boolean {
+		return true;
+	}
+}
